requirement.txt
streamlit==1.50.0
langchain-core==0.3.78
langchain-community==0.3.17
langchain-openai==0.3.0
openai>=1.58.1,<2.0.0
tiktoken==0.11.0
python-dotenv==1.0.1
faiss-cpu==1.12.0
unstructured==0.15.0
beautifulsoup4==4.12.3
requests==2.32.3
python-magic-bin==0.4.14
libmagic==1.0